refactor(admin): extract ContentSettingsPanel from ContentEditor
Moves the editor sidebar — all eight sections (Publish, Ownership,
Bylines, Translations, Taxonomies, SEO, Document Outline, Revisions)
plus the block-panel swap — into ContentSettingsPanel.tsx, along with
the sidebar-only BylineCreditsEditor and AuthorSelector components.

The panel is memoized: ContentEditor re-renders on every keystroke
(formData state), and the panel subtree is expensive. Handler props
are made useCallback-stable at both layers (ContentEditor internals
and the ContentEditPage/ContentNewPage inline arrows in router.tsx)
so the memo actually holds. Verified: zero panel re-renders while
typing and across a full autosave cycle.

No visual or behavioral change.
